WIP -- hooking up the scheduler
