+++ schedule-2d-sat-sim.cabal
@@ -16,9 +16,10 @@ library
   import: warnings
   exposed-modules:
     SatSim.Algo.Greedy
-    SatSim.AMQP
     SatSim.Cache
+    SatSim.Consumer.AMQP
     SatSim.Gen.Producer
+    SatSim.Producer.AMQP
     SatSim.Quantities
     SatSim.Satellite
     SatSim.Schedulable
@@ -54,7 +55,7 @@ executable schedule-2d-sat-sim
     hw-kafka-client,
     optparse-applicative,
     schedule-2d-sat-sim,
-    text
+    text,
 
   hs-source-dirs: app
   default-language: Haskell2010
